rename SearchIndexClient -> SearchClient
