+++ sdk/search/azure-search-documents/azure/search/documents/__init__.py
@@ -33,7 +33,7 @@ from ._index import (
     IndexAction,
     IndexDocumentsBatch,
     IndexingResult,
-    SearchIndexClient,
+    SearchClient,
     SearchQuery,
     SearchItemPaged,
     SuggestQuery,
@@ -112,7 +112,7 @@ __all__ = (
     "IndexAction",
     "IndexDocumentsBatch",
     "IndexingResult",
-    "SearchIndexClient",
+    "SearchClient",
     "SearchItemPaged",
     "SearchQuery",
     "SearchServiceClient",
