+++ sdk/search/azure-search-documents/azure/search/documents/_index/__init__.py
@@ -3,10 +3,10 @@
 # Licensed under the MIT License.
 # ------------------------------------
 from ._index_documents_batch import IndexDocumentsBatch  # pylint: disable=unused-import
-from ._search_index_client import (  # pylint: disable=unused-import
+from ._search_client import (  # pylint: disable=unused-import
     odata,
     SearchItemPaged,
-    SearchIndexClient,
+    SearchClient,
 )
 from ._queries import (  # pylint: disable=unused-import
     AutocompleteQuery,
