+++ sdk/search/azure-search-documents/azure/search/documents/_index/aio/__init__.py
@@ -2,6 +2,6 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 # ------------------------------------
-from ._search_index_client_async import AsyncSearchItemPaged, SearchIndexClient
+from ._search_client_async import AsyncSearchItemPaged, SearchClient
 
-__all__ = ("AsyncSearchItemPaged", "SearchIndexClient")
+__all__ = ("AsyncSearchItemPaged", "SearchClient")
