+++ sdk/search/azure-search-documents/samples/async_samples/sample_authentication_async.py
@@ -26,12 +26,12 @@ import os
 async def authentication_with_api_key_credential_async():
     # [START create_search_client_with_key_async]
     from azure.core.credentials import AzureKeyCredential
-    from azure.search.documents.aio import SearchIndexClient
+    from azure.search.documents.aio import SearchClient
     service_endpoint = os.getenv("AZURE_SEARCH_SERVICE_ENDPOINT")
     index_name = os.getenv("AZURE_SEARCH_INDEX_NAME")
     key = os.getenv("AZURE_SEARCH_API_KEY")
 
-    search_client = SearchIndexClient(service_endpoint, index_name, AzureKeyCredential(key))
+    search_client = SearchClient(service_endpoint, index_name, AzureKeyCredential(key))
     # [END create_search_client_with_key_async]
 
     async with search_client:
