+++ sdk/search/azure-search-documents/samples/async_samples/sample_autocomplete_async.py
@@ -31,10 +31,10 @@ key = os.getenv("AZURE_SEARCH_API_KEY")
 async def autocomplete_query():
     # [START autocomplete_query_async]
     from azure.core.credentials import AzureKeyCredential
-    from azure.search.documents.aio import SearchIndexClient
+    from azure.search.documents.aio import SearchClient
     from azure.search.documents import AutocompleteQuery
 
-    search_client = SearchIndexClient(service_endpoint, index_name, AzureKeyCredential(key))
+    search_client = SearchClient(service_endpoint, index_name, AzureKeyCredential(key))
 
     query = AutocompleteQuery(search_text="bo", suggester_name="sg")
