+++ sdk/search/azure-search-documents/samples/async_samples/sample_crud_operations_async.py
@@ -28,10 +28,10 @@ index_name = os.getenv("AZURE_SEARCH_INDEX_NAME")
 key = os.getenv("AZURE_SEARCH_API_KEY")
 
 from azure.core.credentials import AzureKeyCredential
-from azure.search.documents.aio import SearchIndexClient
+from azure.search.documents.aio import SearchClient
 from azure.search.documents import SearchQuery
 
-search_client = SearchIndexClient(service_endpoint, index_name, AzureKeyCredential(key))
+search_client = SearchClient(service_endpoint, index_name, AzureKeyCredential(key))
 
 async def upload_document():
     # [START upload_document_async]
