+++ sdk/search/azure-search-documents/samples/async_samples/sample_simple_query_async.py
@@ -31,9 +31,9 @@ key = os.getenv("AZURE_SEARCH_API_KEY")
 async def simple_text_query():
     # [START simple_query_async]
     from azure.core.credentials import AzureKeyCredential
-    from azure.search.documents.aio import SearchIndexClient
+    from azure.search.documents.aio import SearchClient
 
-    search_client = SearchIndexClient(service_endpoint, index_name, AzureKeyCredential(key))
+    search_client = SearchClient(service_endpoint, index_name, AzureKeyCredential(key))
 
     async with search_client:
         results = await search_client.search(query="spa")
