+++ sdk/search/azure-search-documents/samples/sample_autocomplete.py
@@ -29,9 +29,9 @@ key = os.getenv("AZURE_SEARCH_API_KEY")
 def autocomplete_query():
     # [START autocomplete_query]
     from azure.core.credentials import AzureKeyCredential
-    from azure.search.documents import AutocompleteQuery, SearchIndexClient
+    from azure.search.documents import AutocompleteQuery, SearchClient
 
-    search_client = SearchIndexClient(service_endpoint, index_name, AzureKeyCredential(key))
+    search_client = SearchClient(service_endpoint, index_name, AzureKeyCredential(key))
 
     query = AutocompleteQuery(search_text="bo", suggester_name="sg")
