+++ sdk/search/azure-search-documents/samples/sample_simple_query.py
@@ -29,9 +29,9 @@ key = os.getenv("AZURE_SEARCH_API_KEY")
 def simple_text_query():
     # [START simple_query]
     from azure.core.credentials import AzureKeyCredential
-    from azure.search.documents import SearchIndexClient
+    from azure.search.documents import SearchClient
 
-    search_client = SearchIndexClient(service_endpoint, index_name, AzureKeyCredential(key))
+    search_client = SearchClient(service_endpoint, index_name, AzureKeyCredential(key))
 
     results = search_client.search(query="spa")
