+++ sdk/search/azure-search-documents/samples/sample_suggestions.py
@@ -29,9 +29,9 @@ key = os.getenv("AZURE_SEARCH_API_KEY")
 def suggest_query():
     # [START suggest_query]
     from azure.core.credentials import AzureKeyCredential
-    from azure.search.documents import SearchIndexClient, SuggestQuery
+    from azure.search.documents import SearchClient, SuggestQuery
 
-    search_client = SearchIndexClient(service_endpoint, index_name, AzureKeyCredential(key))
+    search_client = SearchClient(service_endpoint, index_name, AzureKeyCredential(key))
 
     query = SuggestQuery(search_text="coffee", suggester_name="sg")
