+++ sdk/search/azure-search-documents/tests/async_tests/test_index_live_async.py
@@ -29,7 +29,7 @@ from azure.search.documents import (
     SearchQuery,
     SuggestQuery,
 )
-from azure.search.documents.aio import SearchIndexClient
+from azure.search.documents.aio import SearchClient
 
 TIME_TO_SLEEP = 3
 
@@ -47,13 +47,13 @@ def await_prepared_test(test_fn):
     return run
 
 
-class SearchIndexClientTestAsync(AzureMgmtTestCase):
+class SearchClientTestAsync(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_async_get_document_count(
         self, api_key, endpoint, index_name, **kwargs
     ):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         async with client:
@@ -62,7 +62,7 @@ class SearchIndexClientTestAsync(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_document(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         async with client:
@@ -76,7 +76,7 @@ class SearchIndexClientTestAsync(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_document_missing(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         async with client:
@@ -86,7 +86,7 @@ class SearchIndexClientTestAsync(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_search_simple(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         async with client:
@@ -103,7 +103,7 @@ class SearchIndexClientTestAsync(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_search_filter(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
 
@@ -132,7 +132,7 @@ class SearchIndexClientTestAsync(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_search_counts(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
 
@@ -147,7 +147,7 @@ class SearchIndexClientTestAsync(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_search_coverage(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
 
@@ -166,7 +166,7 @@ class SearchIndexClientTestAsync(AzureMgmtTestCase):
     async def test_get_search_facets_none(
         self, api_key, endpoint, index_name, **kwargs
     ):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
 
@@ -182,7 +182,7 @@ class SearchIndexClientTestAsync(AzureMgmtTestCase):
     async def test_get_search_facets_result(
         self, api_key, endpoint, index_name, **kwargs
     ):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
 
@@ -201,7 +201,7 @@ class SearchIndexClientTestAsync(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_autocomplete(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         async with client:
@@ -212,7 +212,7 @@ class SearchIndexClientTestAsync(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_suggest(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         async with client:
@@ -226,7 +226,7 @@ class SearchIndexClientTestAsync(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_upload_documents_new(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         DOCUMENTS = [
@@ -256,7 +256,7 @@ class SearchIndexClientTestAsync(AzureMgmtTestCase):
     async def test_upload_documents_existing(
         self, api_key, endpoint, index_name, **kwargs
     ):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         DOCUMENTS = [
@@ -273,7 +273,7 @@ class SearchIndexClientTestAsync(AzureMgmtTestCase):
     async def test_delete_documents_existing(
         self, api_key, endpoint, index_name, **kwargs
     ):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         async with client:
@@ -300,7 +300,7 @@ class SearchIndexClientTestAsync(AzureMgmtTestCase):
     async def test_delete_documents_missing(
         self, api_key, endpoint, index_name, **kwargs
     ):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         async with client:
@@ -327,7 +327,7 @@ class SearchIndexClientTestAsync(AzureMgmtTestCase):
     async def test_merge_documents_existing(
         self, api_key, endpoint, index_name, **kwargs
     ):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         async with client:
@@ -354,7 +354,7 @@ class SearchIndexClientTestAsync(AzureMgmtTestCase):
     async def test_merge_documents_missing(
         self, api_key, endpoint, index_name, **kwargs
     ):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         async with client:
@@ -381,7 +381,7 @@ class SearchIndexClientTestAsync(AzureMgmtTestCase):
     async def test_merge_or_upload_documents(
         self, api_key, endpoint, index_name, **kwargs
     ):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         async with client:
