+++ sdk/search/azure-search-documents/tests/test_index_live.py
@@ -22,18 +22,18 @@ from azure.core.exceptions import HttpResponseError
 from azure.core.credentials import AzureKeyCredential
 from azure.search.documents import (
     AutocompleteQuery,
-    SearchIndexClient,
+    SearchClient,
     SearchQuery,
     SuggestQuery,
 )
 
 TIME_TO_SLEEP = 3
 
-class SearchIndexClientTest(AzureMgmtTestCase):
+class SearchClientTest(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     def test_get_document_count(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         assert client.get_document_count() == 10
@@ -41,7 +41,7 @@ class SearchIndexClientTest(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     def test_get_document(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         for hotel_id in range(1, 11):
@@ -54,7 +54,7 @@ class SearchIndexClientTest(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     def test_get_document_missing(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         with pytest.raises(HttpResponseError):
@@ -63,7 +63,7 @@ class SearchIndexClientTest(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     def test_get_search_simple(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         results = list(client.search(query="hotel"))
@@ -75,7 +75,7 @@ class SearchIndexClientTest(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     def test_get_search_filter(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
 
@@ -101,7 +101,7 @@ class SearchIndexClientTest(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     def test_get_search_counts(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
 
@@ -116,7 +116,7 @@ class SearchIndexClientTest(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     def test_get_search_coverage(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
 
@@ -133,7 +133,7 @@ class SearchIndexClientTest(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     def test_get_search_facets_none(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
 
@@ -146,7 +146,7 @@ class SearchIndexClientTest(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     def test_get_search_facets_result(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
 
@@ -164,7 +164,7 @@ class SearchIndexClientTest(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     def test_autocomplete(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         query = AutocompleteQuery(search_text="mot", suggester_name="sg")
@@ -174,7 +174,7 @@ class SearchIndexClientTest(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     def test_suggest(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         query = SuggestQuery(search_text="mot", suggester_name="sg")
@@ -187,7 +187,7 @@ class SearchIndexClientTest(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     def test_upload_documents_new(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         DOCUMENTS = [
@@ -213,7 +213,7 @@ class SearchIndexClientTest(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     def test_upload_documents_existing(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         DOCUMENTS = [
@@ -227,7 +227,7 @@ class SearchIndexClientTest(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     def test_delete_documents_existing(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         results = client.delete_documents([{"hotelId": "3"}, {"hotelId": "4"}])
@@ -249,7 +249,7 @@ class SearchIndexClientTest(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     def test_delete_documents_missing(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         results = client.delete_documents([{"hotelId": "1000"}, {"hotelId": "4"}])
@@ -271,7 +271,7 @@ class SearchIndexClientTest(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     def test_merge_documents_existing(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         results = client.merge_documents(
@@ -295,7 +295,7 @@ class SearchIndexClientTest(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     def test_merge_documents_missing(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         results = client.merge_documents(
@@ -319,7 +319,7 @@ class SearchIndexClientTest(AzureMgmtTestCase):
     @ResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     def test_merge_or_upload_documents(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchIndexClient(
+        client = SearchClient(
             endpoint, index_name, AzureKeyCredential(api_key)
         )
         results = client.merge_or_upload_documents(
